+++ lib/fog/brightbox/requests/compute/destroy_server.rb
@@ -4,6 +4,7 @@ module Fog
       class Real
 
         def destroy_server(identifier, options = {})
+          return nil if identifier.nil? || identifier == ""
           request(
             :expects  => [202],
             :method   => 'DELETE',
